chore: adjust provider to use muxed tf provider

+++ apis/objectstorage/v1alpha1/zz_managedobjectstoragepolicy_types.go
@@ -32,7 +32,7 @@ type ManagedObjectStoragePolicyObservation struct {
 
 	// (Number) Attachment count.
 	// Attachment count.
-	AttachmentCount *float64 `json:"attachmentCount,omitempty" tf:"attachment_count,omitempty"`
+	AttachmentCount *int64 `json:"attachmentCount,omitempty" tf:"attachment_count,omitempty"`
 
 	// (String) Creation time.
 	// Creation time.
@@ -121,8 +121,8 @@ type ManagedObjectStoragePolicyStatus struct {
 // +kubebuilder:storageversion
 
 // ManagedObjectStoragePolicy is the Schema for the ManagedObjectStoragePolicys API. This resource represents an UpCloud Managed Object Storage policy.
-// +kubebuilder:printcolumn:name="READY",type="string",JSONPath=".status.conditions[?(@.type=='Ready')].status"
 // +kubebuilder:printcolumn:name="SYNCED",type="string",JSONPath=".status.conditions[?(@.type=='Synced')].status"
+// +kubebuilder:printcolumn:name="READY",type="string",JSONPath=".status.conditions[?(@.type=='Ready')].status"
 // +kubebuilder:printcolumn:name="EXTERNAL-NAME",type="string",JSONPath=".metadata.annotations.crossplane\\.io/external-name"
 // +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
 // +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,upcloud}
